All other changes to books module so the book controller changes make sense

+++ server-side/src/modules/books/book.services.ts
@@ -1,41 +1,10 @@
-import { createBookInput, updateBookInput } from "./book.schemas";
 import Book from "./book.model";
-import Chapter from "../chapters/chapter.model";
 import { deleteAllChaptersFromBook } from "../chapters/chapter.services";
-
-interface addBookData extends createBookInput {
-  userId: string;
-}
-
-export async function addBookToDatabase({
-  title,
-  description,
-  userId,
-}: addBookData) {
-  return await Book.create({
-    title,
-    description,
-    author: userId,
-  });
-}
-
-export async function updateBookInDatabase(
-  bookId: string,
-  updates: updateBookInput
-) {
-  return await Book.findByIdAndUpdate(bookId, updates);
-}
-
-export async function getAllBooksFromUser(userId: string) {
-  return await Book.find({ author: userId });
-}
-
-export async function getAllBooksByIdFromDB(bookId: string) {
-  return await Book.findById(bookId);
-}
+import { HttpError } from "../../errors/httpError";
 
 export async function deleteBookAndChapters(bookId: string) {
-  await Book.findByIdAndDelete(bookId);
+  const book = await Book.findByIdAndDelete(bookId);
+  if (!book) throw new HttpError(404, "Not found");
   await deleteAllChaptersFromBook(bookId);
   return;
 }
